requirements.txt
praw
pandas
matplotlib.pyplot
seaborn
collections